.ipynb_checkpoints/requirements-checkpoint.txt
pandas==1.4.2
numpy==1.21.5
torchvision==0.14.1
transformers==4.28.1
streamlit==1.21.0